requirements.txt
kafka-python==2.2.15
confluent-kafka>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
